refactor: split allowance controller (#24)
