+++ api/controllers/allowance_controller.go
@@ -1,38 +1,1 @@
 package controllers
-
-import (
-	"github.com/gin-gonic/gin"
-	"github.com/t2469/labor-management-system.git/db"
-	"github.com/t2469/labor-management-system.git/models"
-	"net/http"
-)
-
-func CrateAllowanceType(c *gin.Context) {
-	var at models.AllowanceType
-	if err := c.ShouldBindJSON(&at); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	if err := db.DB.Create(&at).Error; err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	c.JSON(http.StatusOK, at)
-}
-
-func CreateEmployeeAllowance(c *gin.Context) {
-	var ea models.EmployeeAllowance
-	if err := c.ShouldBindJSON(&ea); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	if err := db.DB.Create(&ea).Error; err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	c.JSON(http.StatusOK, ea)
-}
